requirements.txt
pyserial>=3.5
python-can>=4.2.0
dataclasses-json>=0.5.7
pydantic>=2.0.0
loguru>=0.7.0